TfvcRootChecker: add tests for server workspace (failing yet)
